refactor: use class-based approach for module
Switch to using a class-based approach, rather than loose-functions.
This makes more sense as there is data, such as start time and timers
which needs to be associated together, which is the whole point
of a class. Also refactored the time generation function into several
smaller ones to make it easier to read and maintain. Cleaned up the
tests and added some more as well.

Also started generating the documentation from JSDoc, rather than
trying to do so by hand (as I did with _gulp-errorable_ and
_terminal-spawn_). This way the documentation will be located
closer to the code than if it only exists in a separate readme
somewhere. It will also be visible on-hover in the editor. Meaning
that it will be much more likely to be correct, than the manual
approach. However, it is still not generated from the code
as I couldn't get typedoc to work, due to a bug in the markdown
generation module.

Signed-off-by: David Piper <david.piper@mailfence.com>