feat: make more generic and extensible
As part of ongoing work on getting the module ready to publish,
I have made it much more generic. This is part of work to
allow the user to specify their own messages, rather than
being forced to use the default one that I personally use.

To this end the following things were done:

- added options object to constructor
- allow passing in success or failure messages as strings
- allow passing in success or failure messages as functions which
  will be lazily evaluated
- added an `includeTime` option, which is not currently used, but
  will enable or disable the time part of the message if desired
- added a `scriptName` option which allows the user to specify the
  name of their script rather than being hard-coded to "Verification".
- added an instance function to get the elapsed time: `getElapsedTime`
  which will allow the user to still incorporate the elapsed time, as
  tracked by `slack-notify-status` into their messages

Each of these pieces works together to help facilitate the genericity
of the module, and of the messages in particular. It makes sense from
a UX perspective to allow the user to customize the message as much
as they want, since the core responsibility of the module is to simply
send slack messages to the designated endpoint. In fact, it might
actually be a good idea to split out the message generation part
of the module into a separate module...

Signed-off-by: David Piper <david.piper@mailfence.com>