feat: use timer module instead of doing it here
Use the new @dbpiper/timer module that I just created to do the
timing of the code, rather than trying to do it in this module.
This significantly cleans up the code and makes it much less
convoluted, by decoupling the timing logic from the sending
logic.

Also make the messages not configurable, I think this was trying to do
too much or "over-engineering" the module. The module only exists
for my own personal use at the moment, there might be a point in the
future when someone else uses it, however trying to anticipate that
and what a theoretical person might want seems futile. Instead, it
simply takes in the name of the script which **is** customizable
and the timer, which is @dbpiper/timer by default.

The timer is assumed to have the same API as @dbpiper/timer, so
_technically_ it will work with any timer, just as long as that timer
behaves nearly identically to @dbpiper/timer...

Signed-off-by: David Piper <david.piper@mailfence.com>