docs: write readme, remove jsdoc generated one
The JSDoc readme generation idea doesn't seem to work nearly
as well as I had hoped I would. It generates dry documentation
which doesn't have enough descriptions and the layout is very
hard to read. In other words, its much better to just write
the readme the old-school way for now.

I'm hopeful that in the future something like typedoc will get
good and that may be able to augment this process a bit, but
I'm not very confident that this will happen anytime soon.

For now though, I have written the readme the standard way. I also
added badges header and the license footer. I also changed the
rejected promise value to simply contain the error, since it
isn't very helpful to know what the message is if we can't even
send it!

Signed-off-by: David Piper <david.piper@mailfence.com>