feat: expose the underlying timer so that it can be started manually
Until now the timer only ever started when the instance object was
created, however it is often convenient to create the instance
object early and then start the timer later on. This adds support
for this, by adding an accessor to underlying timer, in other
words the user can now call slackNotifyStatus.timer.start()
whenever they wish.

Signed-off-by: David Piper <david.piper@mailfence.com>